% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{fqtk_demux}
\alias{fqtk_demux}
\title{Function to call the fqtk CLI command for demultiplexing.
Exposes the demux functionality as a Rust function that can be called from R.}
\usage{
fqtk_demux(inputs, max_mismatches, read_structures, sample_metadata, output)
}
\description{
Function to call the fqtk CLI command for demultiplexing.
Exposes the demux functionality as a Rust function that can be called from R.
}
